- fixed several scale related floating point conversion warnings.

+++ src/gamedata/d_dehacked.cpp
@@ -1193,7 +1193,7 @@ static int PatchThing (int thingy)
 			}
 			else if (stricmp (Line1, "Scale") == 0)
 			{
-				info->Scale.Y = info->Scale.X = clamp(atof (Line2), 1./65536, 256.);
+				info->Scale.Y = info->Scale.X = clamp((float)atof (Line2), 1.f/65536, 256.f);
 			}
 			else if (stricmp (Line1, "Decal") == 0)
 			{
